pyproject.toml
[project]
name = "macrotrade-pyspark"
version = "0.1.0"
description = "data transformations"
authors = [
    {name = "Yanis", email = "konstantinidisj@gmail.com"}
]
license = "Apache-2.0"
dependencies = [
    "numpy>=1.20.2,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "pyarrow>=12.0.0,<13.0.0",
    "requests>=2.32.3,<3.0.0",
    "comtradeapicall>=1.2.1,<2.0.0",
    "duckdb>=1.2.2,<2.0.0",
    "pyspark==3.5.1",
    "apache-airflow-providers-apache-spark>=5.2.1",
    "apache-airflow>=3.0.0",
    "azure-identity>=1.22.0",
    "azure-storage-blob>=12.25.1",
]
requires-python = ">=3.11,<3.12"

[project.optional-dependencies]
dev = [
    "pylint>=2.8.2,<3.0.0",
    "pytest>=6.2.4,<7.0.0",
    "mypy>=0.990,<1.0.0"
]

[tool.setuptools]
packages = ["data_transformations", "jobs", "resources"]
package-dir = {"" = "."}

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
